Converted Quad4 master-elements to use Kokkos views. (#210)
* Converted Quad4 master-elements to use Kokkos views.

* Resolve merge conflicts....
