+++ include/KernelBuilder.h
@@ -114,6 +114,7 @@ namespace nalu{
 
     bool isNotNGP = !(topo == stk::topology::HEXAHEDRON_8 ||
                       topo == stk::topology::HEXAHEDRON_27 ||
+                      topo == stk::topology::QUADRILATERAL_4_2D ||
                       topo == stk::topology::WEDGE_6 ||
                       topo == stk::topology::TETRAHEDRON_4 ||
                       topo == stk::topology::PYRAMID_5);
