+++ src/element_promotion/PromoteElement.C
@@ -10,6 +10,7 @@
 #include <element_promotion/ElementDescription.h>
 #include <master_element/Hex8CVFEM.h>
 #include <master_element/MasterElement.h>
+#include <master_element/Quad42DCVFEM.h>  
 #include <NaluEnv.h>
 #include <BucketLoop.h>
 
@@ -661,7 +662,7 @@ set_coordinates_quad(
   const stk::mesh::PartVector& promotedPartVector,
   const VectorFieldType& coordField)
 {
-  Quad2DSCS meQuad;
+  Quad42DSCS meQuad;
   auto selector = stk::mesh::selectUnion(promotedPartVector);
   const auto& elem_buckets = bulk.get_buckets(stk::topology::ELEM_RANK, selector);
