+++ src/master_element/MasterElement.C
@@ -89,708 +89,6 @@ MasterElement::vector_norm_sq( const double * vect, int len )
 }
 
 //--------------------------------------------------------------------------
-//-------- constructor -----------------------------------------------------
-//--------------------------------------------------------------------------
-Quad2DSCV::Quad2DSCV()
-  : MasterElement()
-{
-  nDim_ = 2;
-  nodesPerElement_ = 4;
-  numIntPoints_ = 4;
-
-  // define ip node mappings
-  ipNodeMap_.resize(4);
-  ipNodeMap_[0] = 0; ipNodeMap_[1] = 1; ipNodeMap_[2] = 2; ipNodeMap_[3] = 3;
-
-  // standard integration location
-  intgLoc_.resize(8);    
-  intgLoc_[0]  = -0.25; intgLoc_[1]  = -0.25; 
-  intgLoc_[2]  = +0.25; intgLoc_[3]  = -0.25; 
-  intgLoc_[4]  = +0.25; intgLoc_[5]  = +0.25; 
-  intgLoc_[6]  = -0.25; intgLoc_[7]  = +0.25; 
-
-  // shifted integration location
-  intgLocShift_.resize(8);    
-  intgLocShift_[0]  = -0.50; intgLocShift_[1]  = -0.50; 
-  intgLocShift_[2]  = +0.50; intgLocShift_[3]  = -0.50; 
-  intgLocShift_[4]  = +0.50; intgLocShift_[5]  = +0.50; 
-  intgLocShift_[6]  = -0.50; intgLocShift_[7]  = +0.50; 
-}
-
-//--------------------------------------------------------------------------
-//-------- destructor ------------------------------------------------------
-//--------------------------------------------------------------------------
-Quad2DSCV::~Quad2DSCV()
-{
-  // does nothing
-}
-
-//--------------------------------------------------------------------------
-//-------- ipNodeMap -------------------------------------------------------
-//--------------------------------------------------------------------------
-const int *
-Quad2DSCV::ipNodeMap(
-  int /*ordinal*/)
-{
-  // define scv->node mappings
-  return &ipNodeMap_[0];
-}
-
-//--------------------------------------------------------------------------
-//-------- determinant -----------------------------------------------------
-//--------------------------------------------------------------------------
-void Quad2DSCV::determinant(
-  const int nelem,
-  const double *coords,
-  double *volume,
-  double *error)
-{
-  int lerr = 0;
-
-  SIERRA_FORTRAN(quad_scv_det)
-    ( &nelem, &nodesPerElement_, &numIntPoints_, coords,
-      volume, error, &lerr );
-}
-
-//--------------------------------------------------------------------------
-//-------- shape_fcn -------------------------------------------------------
-//--------------------------------------------------------------------------
-void
-Quad2DSCV::shape_fcn(double *shpfc)
-{
-  quad_shape_fcn(numIntPoints_, &intgLoc_[0], shpfc);
-}
-
-//--------------------------------------------------------------------------
-//-------- shifted_shape_fcn -----------------------------------------------
-//--------------------------------------------------------------------------
-void
-Quad2DSCV::shifted_shape_fcn(double *shpfc)
-{
-  quad_shape_fcn(numIntPoints_, &intgLocShift_[0], shpfc);
-}
-
-//--------------------------------------------------------------------------
-//-------- quad_shape_fcn ---------------------------------------------------
-//--------------------------------------------------------------------------
-void
-Quad2DSCV::quad_shape_fcn(
-  const int  &npts,
-  const double *isoParCoord, 
-  double *shape_fcn)
-{
-  for (int j = 0; j < npts; ++j ) {
-    const int fourj = 4*j;
-    const int k = 2*j;
-    const double s1 = isoParCoord[k];
-    const double s2 = isoParCoord[k+1];
-    shape_fcn[    fourj] = 1.0/4.0 + 0.5*(-s1 - s2 ) + s1*s2;
-    shape_fcn[1 + fourj] = 1.0/4.0 + 0.5*( s1 - s2 ) - s1*s2;
-    shape_fcn[2 + fourj] = 1.0/4.0 + 0.5*( s1 + s2 ) + s1*s2;
-    shape_fcn[3 + fourj] = 1.0/4.0 + 0.5*(-s1 + s2 ) - s1*s2;
-  }
-}
-
-//--------------------------------------------------------------------------
-//-------- constructor -----------------------------------------------------
-//--------------------------------------------------------------------------
-Quad2DSCS::Quad2DSCS()
-  : MasterElement()
-{
-  nDim_ = 2;
-  nodesPerElement_ = 4;
-  numIntPoints_ = 4;
-  scaleToStandardIsoFac_ = 2.0;
-
-  // define L/R mappings
-  lrscv_.resize(8);
-  lrscv_[0]  = 0; lrscv_[1]  = 1;
-  lrscv_[2]  = 1; lrscv_[3]  = 2;
-  lrscv_[4]  = 2; lrscv_[5]  = 3;
-  lrscv_[6]  = 0; lrscv_[7]  = 3;
-  
-  // define opposing node
-  oppNode_.resize(8);
-  // face 0; nodes 0,1
-  oppNode_[0] = 3; oppNode_[1] = 2;
-  // face 1; nodes 1,2
-  oppNode_[2] = 0; oppNode_[3] = 3;
-  // face 2; nodes 2,3
-  oppNode_[4] = 1; oppNode_[5] = 0;
-  // face 3; nodes 3,0
-  oppNode_[6] = 2; oppNode_[7] = 1;
-
-  // define opposing face
-  oppFace_.resize(8);
-  // face 0
-  oppFace_[0]  = 3; oppFace_[1] = 1;
-  // face 1
-  oppFace_[2]  = 0; oppFace_[3] = 2;
-  // face 2
-  oppFace_[4]  = 1; oppFace_[5] = 3;  
-  // face 3
-  oppFace_[6]  = 2; oppFace_[7] = 0;
-
-  // standard integration location
-  intgLoc_.resize(8);    
-  intgLoc_[0] =  0.00; intgLoc_[1] = -0.25; // surf 1; 1->2
-  intgLoc_[2] =  0.25; intgLoc_[3] =  0.00; // surf 2; 2->3
-  intgLoc_[4] =  0.00; intgLoc_[5] =  0.25; // surf 3; 3->4
-  intgLoc_[6] = -0.25; intgLoc_[7] =  0.00; // surf 3; 1->5
-
-  // shifted
-  intgLocShift_.resize(8);
-  intgLocShift_[0] =  0.00; intgLocShift_[1] = -0.50;
-  intgLocShift_[2] =  0.50; intgLocShift_[3] =  0.00;
-  intgLocShift_[4] =  0.00; intgLocShift_[5] =  0.50;
-  intgLocShift_[6] = -0.50; intgLocShift_[7] =  0.00;
-
-  // exposed face
-  intgExpFace_.resize(16);
-  // face 0; scs 0, 1; nodes 0,1
-  intgExpFace_[0]  = -0.25; intgExpFace_[1]  = -0.50; 
-  intgExpFace_[2]  =  0.25; intgExpFace_[3]  = -0.50;
-  // face 1; scs 0, 1; nodes 1,2
-  intgExpFace_[4]  =  0.50; intgExpFace_[5]  = -0.25;
-  intgExpFace_[6]  =  0.50; intgExpFace_[7]  =  0.25;
-  // face 2, surf 0, 1; nodes 2,3
-  intgExpFace_[8]  =  0.25; intgExpFace_[9]  =  0.50;
-  intgExpFace_[10] = -0.25; intgExpFace_[11] =  0.50;
-  // face 3, surf 0, 1; nodes 3,0
-  intgExpFace_[12] = -0.50; intgExpFace_[13] =  0.25;
-  intgExpFace_[14] = -0.50; intgExpFace_[15] = -0.25; 
-
-  // boundary integration point ip node mapping (ip on an ordinal to local node number)
-  ipNodeMap_.resize(8); // 2 ips * 4 faces
-  // face 0;
-  ipNodeMap_[0] = 0;  ipNodeMap_[1] = 1;  
-  // face 1;
-  ipNodeMap_[2] = 1;  ipNodeMap_[3] = 2; 
-  // face 2; 
-  ipNodeMap_[4] = 2;  ipNodeMap_[5] = 3;  
-  // face 3;
-  ipNodeMap_[6] = 3;  ipNodeMap_[7] = 0; 
-
-  sideNodeOrdinals_ = {
-      0, 1,
-      1, 2,
-      2, 3,
-      3, 0
-  };
-
-  std::vector<std::vector<double>> nodeLocations =
-  {
-      {-0.5,-0.5}, {+0.5,-0.5},
-      {+0.5,+0.5}, {-0.5,+0.5}
-  };
-  intgExpFaceShift_.resize(16);
-  int index = 0;
-  stk::topology topo = stk::topology::QUADRILATERAL_4_2D;
-  for (unsigned k = 0; k < topo.num_sides(); ++k) {
-    stk::topology side_topo = topo.side_topology(k);
-    const int* ordinals = side_node_ordinals(k);
-    for (unsigned n = 0; n < side_topo.num_nodes(); ++n) {
-      intgExpFaceShift_[2*index + 0] = nodeLocations[ordinals[n]][0];
-      intgExpFaceShift_[2*index + 1] = nodeLocations[ordinals[n]][1];
-      ++index;
-    }
-  }
-}
-
-//--------------------------------------------------------------------------
-//-------- destructor ------------------------------------------------------
-//--------------------------------------------------------------------------
-Quad2DSCS::~Quad2DSCS()
-{
-  // does nothing
-}
-
-//--------------------------------------------------------------------------
-//-------- ipNodeMap -------------------------------------------------------
-//--------------------------------------------------------------------------
-const int *
-Quad2DSCS::ipNodeMap(
-  int ordinal)
-{
-  // define ip->node mappings for each face (ordinal); 
-  return &ipNodeMap_[ordinal*2];
-}
-
-
-//--------------------------------------------------------------------------
-//-------- side_node_ordinals ----------------------------------------------
-//--------------------------------------------------------------------------
-const int *
-Quad2DSCS::side_node_ordinals(
-  int ordinal)
-{
-  // define face_ordinal->node_ordinal mappings for each face (ordinal);
-  return &sideNodeOrdinals_[ordinal*2];
-}
-
-//--------------------------------------------------------------------------
-//-------- determinant -----------------------------------------------------
-//--------------------------------------------------------------------------
-void Quad2DSCS::determinant(
-  const int nelem,
-  const double *coords,
-  double *areav,
-  double *error)
-{
-  SIERRA_FORTRAN(quad_scs_det)
-    ( &nelem, &nodesPerElement_, &numIntPoints_, coords, areav );
-
-  // all is always well; no error checking
-  *error = 0;
-}
-
-//--------------------------------------------------------------------------
-//-------- grad_op ---------------------------------------------------------
-//--------------------------------------------------------------------------
-void Quad2DSCS::grad_op(
-  const int nelem,
-  const double *coords,
-  double *gradop,
-  double *deriv,
-  double *det_j,
-  double *error)
-{
-  int lerr = 0;
-
-  SIERRA_FORTRAN(quad_derivative)
-    ( &numIntPoints_, &intgLoc_[0], deriv );
-  
-  SIERRA_FORTRAN(quad_gradient_operator)
-    ( &nelem,
-      &nodesPerElement_,
-      &numIntPoints_,
-      deriv,
-      coords, gradop, det_j, error, &lerr );
-  
-  if ( lerr )
-    std::cout << "sorry, negative Quad2DSCS volume.." << std::endl;  
-}
-
-//--------------------------------------------------------------------------
-//-------- shifted_grad_op -------------------------------------------------
-//--------------------------------------------------------------------------
-void Quad2DSCS::shifted_grad_op(
-  const int nelem,
-  const double *coords,
-  double *gradop,
-  double *deriv,
-  double *det_j,
-  double *error)
-{
-  int lerr = 0;
-
-  SIERRA_FORTRAN(quad_derivative)
-    ( &numIntPoints_, &intgLocShift_[0], deriv );
-
-  SIERRA_FORTRAN(quad_gradient_operator)
-    ( &nelem,
-      &nodesPerElement_,
-      &numIntPoints_,
-      deriv,
-      coords, gradop, det_j, error, &lerr );
-
-  if ( lerr )
-    std::cout << "sorry, negative Quad2DSCS volume.." << std::endl;
-}
-
-//--------------------------------------------------------------------------
-//-------- face_grad_op ----------------------------------------------------
-//--------------------------------------------------------------------------
-void Quad2DSCS::face_grad_op(
-  const int nelem,
-  const int face_ordinal,
-  const double *coords,
-  double *gradop,
-  double *det_j,
-  double *error)
-{
-  int lerr = 0;
-  int npf = 2;
-  int ndim = 2;
-
-  const int nface = 1;
-  double dpsi[8];
-
-  for ( int n=0; n<nelem; n++ ) {
-    
-    for ( int k=0; k<npf; k++ ) {
-      
-      const int row = 4*face_ordinal + k*ndim;
-      SIERRA_FORTRAN(quad_derivative)
-        ( &nface, &intgExpFace_[row], dpsi );
-      
-      SIERRA_FORTRAN(quad_gradient_operator)
-        ( &nface,
-          &nodesPerElement_,
-          &nface,
-          dpsi,
-          &coords[8*n], &gradop[k*nelem*8+n*8], &det_j[npf*n+k], error, &lerr );
-      
-      if ( lerr )
-        std::cout << "sorry, issue with face_grad_op.." << std::endl;
-      
-    }
-  }
-}
-
-//--------------------------------------------------------------------------
-//-------- shifted_face_grad_op --------------------------------------------
-//--------------------------------------------------------------------------
-void Quad2DSCS::shifted_face_grad_op(
-  const int nelem,
-  const int face_ordinal,
-  const double *coords,
-  double *gradop,
-  double *det_j,
-  double *error)
-{
-  int lerr = 0;
-  int npf = 2;
-  int ndim = 2;
-
-  const int nface = 1;
-  double dpsi[8];
-
-  for ( int n=0; n<nelem; n++ ) {
-
-    for ( int k=0; k<npf; k++ ) {
-
-      const int row = 4*face_ordinal + k*ndim;
-      SIERRA_FORTRAN(quad_derivative)
-        ( &nface, &intgExpFaceShift_[row], dpsi );
-
-      SIERRA_FORTRAN(quad_gradient_operator)
-        ( &nface,
-          &nodesPerElement_,
-          &nface,
-          dpsi,
-          &coords[8*n], &gradop[k*nelem*8+n*8], &det_j[npf*n+k], error, &lerr );
-
-      if ( lerr )
-        std::cout << "sorry, issue with face_grad_op.." << std::endl;
-
-    }
-  }
-}
-
-//--------------------------------------------------------------------------
-//-------- gij -------------------------------------------------------------
-//--------------------------------------------------------------------------
-void Quad2DSCS::gij(
-  const double *coords,
-  double *gupperij,
-  double *glowerij,
-  double *deriv)
-{
-  SIERRA_FORTRAN(twod_gij)
-    ( &nodesPerElement_,
-      &numIntPoints_,
-      deriv,
-      coords, gupperij, glowerij);
-}
-
-//--------------------------------------------------------------------------
-//-------- adjacentNodes ---------------------------------------------------
-//--------------------------------------------------------------------------
-const int *
-Quad2DSCS::adjacentNodes()
-{
-  // define L/R mappings
-  return &lrscv_[0];
-}
-
-//--------------------------------------------------------------------------
-//-------- opposingNodes --------------------------------------------------
-//--------------------------------------------------------------------------
-int
-Quad2DSCS::opposingNodes(
-  const int ordinal,
-  const int node)
-{
-  return oppNode_[ordinal*2+node];
-}
-
-//--------------------------------------------------------------------------
-//-------- opposingFace --------------------------------------------------
-//--------------------------------------------------------------------------
-int
-Quad2DSCS::opposingFace(
-  const int ordinal,
-  const int node)
-{
-  return oppFace_[ordinal*2+node];
-}
-
-//--------------------------------------------------------------------------
-//-------- shape_fcn -------------------------------------------------------
-//--------------------------------------------------------------------------
-void
-Quad2DSCS::shape_fcn(double *shpfc)
-{
-  quad_shape_fcn(numIntPoints_, &intgLoc_[0], shpfc);
-}
-
-//--------------------------------------------------------------------------
-//-------- shifted_shape_fcn -----------------------------------------------
-//--------------------------------------------------------------------------
-void
-Quad2DSCS::shifted_shape_fcn(double *shpfc)
-{
-  quad_shape_fcn(numIntPoints_, &intgLocShift_[0], shpfc);
-}
-
-//--------------------------------------------------------------------------
-//-------- quad_shape_fcn ---------------------------------------------------
-//--------------------------------------------------------------------------
-void
-Quad2DSCS::quad_shape_fcn(
-  const int  &npts,
-  const double *isoParCoord, 
-  double *shape_fcn)
-{
-  for (int j = 0; j < npts; ++j ) {
-    const int fourj = 4*j;
-    const int k = 2*j;
-    const double s1 = isoParCoord[k];
-    const double s2 = isoParCoord[k+1];
-    shape_fcn[    fourj] = 1.0/4.0 + 0.5*(-s1 - s2 ) + s1*s2;
-    shape_fcn[1 + fourj] = 1.0/4.0 + 0.5*( s1 - s2 ) - s1*s2;
-    shape_fcn[2 + fourj] = 1.0/4.0 + 0.5*( s1 + s2 ) + s1*s2;
-    shape_fcn[3 + fourj] = 1.0/4.0 + 0.5*(-s1 + s2 ) - s1*s2;
-  }
-}
-
-//--------------------------------------------------------------------------
-//-------- isInElement -----------------------------------------------------
-//--------------------------------------------------------------------------
-double
-Quad2DSCS::isInElement(
-  const double *elemNodalCoord,
-  const double *pointCoord,
-  double *isoParCoord )
-{
-  // square of the desired norm, 1.0e-8
-  const double isInElemConverged = 1.0e-16;
-  const int maxNonlinearIter = 10;
-
-  // -1:1 isoparametric range
-  
-  // Translate element so that (x,y) coordinates of the first node are (0,0)
-  double x[4] = {0.,
-                 elemNodalCoord[1] - elemNodalCoord[0],
-                 elemNodalCoord[2] - elemNodalCoord[0],
-                 elemNodalCoord[3] - elemNodalCoord[0] };
-  double y[4] = {0.,
-                 elemNodalCoord[5] - elemNodalCoord[4],
-                 elemNodalCoord[6] - elemNodalCoord[4],
-                 elemNodalCoord[7] - elemNodalCoord[4] };
-  
-  // (xp,yp) is the point at which we're searching for (xi,eta)
-  // (must translate this also)
-  
-  double xp = pointCoord[0] - elemNodalCoord[0];
-  double yp = pointCoord[1] - elemNodalCoord[4];
-  
-  // Newton-Raphson iteration for (xi,eta)
-  double j[4];
-  double f[2];
-  double shapefct[4];
-  
-  double xinew = 0.5;     // initial guess
-  double etanew = 0.5;
-  
-  double xicur = 0.5;
-  double etacur = 0.5;
-  
-  double xidiff[2] = { 1.0, 1.0};
-  int i = 0;
-  
-  bool converged = false;
-  
-  do {
-    xicur = xinew;
-    etacur = etanew;
-
-    j[0]=  0.25*(1.00-etacur)*x[1]
-	  +0.25*(1.00+etacur)*x[2]
-	  -0.25*(1.00+etacur)*x[3];
-
-    j[1]= -0.25*(1.00+xicur)*x[1]
-	  +0.25*(1.00+xicur)*x[2]
-	  +0.25*(1.00-xicur)*x[3];
-
-    j[2]=  0.25*(1.00-etacur)*y[1]
-	  +0.25*(1.00+etacur)*y[2]
-	  -0.25*(1.00+etacur)*y[3];
-
-    j[3]= -0.25*(1.00+xicur)*y[1]
-	  +0.25*(1.00+xicur)*y[2]
-	  +0.25*(1.00-xicur)*y[3];
-
-    double jdet = j[0]*j[3] - j[1]*j[2];
-
-    shapefct[0]=0.25*(1.00-etacur)*(1.00-xicur);
-    shapefct[1]=0.25*(1.00-etacur)*(1.00+xicur);
-    shapefct[2]=0.25*(1.00+etacur)*(1.00+xicur);
-    shapefct[3]=0.25*(1.00+etacur)*(1.00-xicur);
-
-
-    f[0] = (shapefct[1]*x[1]+shapefct[2]*x[2]+shapefct[3]*x[3]) - xp;
-    f[1] = (shapefct[1]*y[1]+shapefct[2]*y[2]+shapefct[3]*y[3]) - yp;
-
-    xinew  = xicur  - ( f[0]*j[3] - f[1]*j[1])/jdet;
-    etanew = etacur - (-f[0]*j[2] + f[1]*j[0])/jdet;
-
-    xidiff[0] = xinew  - xicur;
-    xidiff[1] = etanew - etacur;
-    
-    double vectorNorm = xidiff[0]*xidiff[0] + xidiff[1]*xidiff[1];
-    converged = (vectorNorm < isInElemConverged);
-  }  while ( !converged && (++i < maxNonlinearIter) );
-
-  // set a bad value
-  isoParCoord[0] = isoParCoord[1] = 1.0e6;
-  double dist = 1.0e6;
-  if ( i < maxNonlinearIter ) {
-    isoParCoord[0] = xinew;
-    isoParCoord[1] = etanew;
-    dist = (std::abs(xinew) > std::abs(etanew))
-      ? std::abs(xinew) : std::abs(etanew);
-  }
-  return dist;
-}
-
-//--------------------------------------------------------------------------
-//-------- interpolatePoint ------------------------------------------------
-//--------------------------------------------------------------------------
-void
-Quad2DSCS::interpolatePoint(
-  const int &nComp,
-  const double *isoParCoord,
-  const double *field,
-  double *result )
-{
-  // -1:1 isoparametric range
-  const double xi   = isoParCoord[0];
-  const double eta  = isoParCoord[1];
-
-  for ( int i = 0; i < nComp; i++ )
-  {
-    // Base 'field array' index for ith component
-    int b = 4*i;
-
-    result[i] = 0.25 * (
-      (1-eta) * (1-xi) * field[b+0] +
-      (1-eta) * (1+xi) * field[b+1] +
-      (1+eta) * (1+xi) * field[b+2] +
-      (1+eta) * (1-xi) * field[b+3] ) ;
-  }  
-}
-
-//--------------------------------------------------------------------------
-//-------- general_shape_fcn -----------------------------------------------
-//--------------------------------------------------------------------------
-void
-Quad2DSCS::general_shape_fcn(
-  const int numIp,
-  const double *isoParCoord,
-  double *shpfc)
-{
-  // -1:1 isoparametric range
-  const double npe = nodesPerElement_;
-  for ( int ip = 0; ip < numIp; ++ip ) {
-    
-    const int rowIpc = 2*ip;
-    const int rowSfc = npe*ip;
-    
-    const double s1 = isoParCoord[rowIpc];
-    const double s2 = isoParCoord[rowIpc+1];
-    shpfc[rowSfc  ] = 0.25*(1.0-s1)*(1.0-s2);
-    shpfc[rowSfc+1] = 0.25*(1.0+s1)*(1.0-s2);
-    shpfc[rowSfc+2] = 0.25*(1.0+s1)*(1.0+s2);
-    shpfc[rowSfc+3] = 0.25*(1.0-s1)*(1.0+s2);
-    
-  }
-}
-
-//--------------------------------------------------------------------------
-//-------- general_face_grad_op --------------------------------------------
-//--------------------------------------------------------------------------
-void 
-Quad2DSCS::general_face_grad_op(
-  const int face_ordinal,
-  const double *isoParCoord,
-  const double *coords,
-  double *gradop,
-  double *det_j,
-  double *error)
-{
-  int lerr = 0;
-  const int nface = 1;
-
-  double dpsi[8];
-
-  SIERRA_FORTRAN(quad_derivative)
-    ( &nface, &isoParCoord[0], dpsi );
-      
-  SIERRA_FORTRAN(quad_gradient_operator)
-    ( &nface,
-      &nodesPerElement_,
-      &nface,
-      dpsi,
-      &coords[0], &gradop[0], &det_j[0], error, &lerr );
-  
-  if ( lerr )
-    std::cout << "Quad2DSCS::general_face_grad_op: issue.." << std::endl;
-  
-}
-
-//--------------------------------------------------------------------------
-//-------- sidePcoords_to_elemPcoords --------------------------------------
-//--------------------------------------------------------------------------
-void 
-Quad2DSCS::sidePcoords_to_elemPcoords(
-  const int & side_ordinal,
-  const int & npoints,
-  const double *side_pcoords,
-  double *elem_pcoords)
-{
-  switch (side_ordinal) {
-  case 0:
-    for (int i=0; i<npoints; i++) {
-      elem_pcoords[i*2+0] = 0.5*side_pcoords[i];
-      elem_pcoords[i*2+1] = -0.5;
-    }
-    break;
-  case 1:
-    for (int i=0; i<npoints; i++) {
-      elem_pcoords[i*2+0] = 0.5;
-      elem_pcoords[i*2+1] = 0.5*side_pcoords[i];
-    }
-    break;
-  case 2:
-    for (int i=0; i<npoints; i++) {
-      elem_pcoords[i*2+0] = -0.5*side_pcoords[i];
-      elem_pcoords[i*2+1] = 0.5;
-    }
-    break;
-  case 3:
-    for (int i=0; i<npoints; i++) {
-      elem_pcoords[i*2+0] = -0.5;
-      elem_pcoords[i*2+1] = -0.5*side_pcoords[i];
-    }
-    break;
-  default:
-    throw std::runtime_error("Quad2DSCS::sideMap invalid ordinal");
-  }
-}
-
-//--------------------------------------------------------------------------
 //-------- constructor------------------------------------------------------
 //--------------------------------------------------------------------------
 QuadrilateralP2Element::QuadrilateralP2Element()
