+++ unit_tests/UnitTestKokkosME.C
@@ -210,6 +210,17 @@ TEST(KokkosME, test_tet4_views)
   );
 }
 
+TEST(KokkosME, test_quad42D_views)
+{
+  test_ME_views<sierra::nalu::AlgTraitsQuad4_2D>(
+    {
+     sierra::nalu::SCS_AREAV,
+     sierra::nalu::SCS_GRAD_OP,
+     sierra::nalu::SCV_VOLUME
+    }
+  );
+}
+
 TEST(KokkosME, test_wed6_views)
 {
   test_ME_views<sierra::nalu::AlgTraitsWed6>(
@@ -236,11 +247,30 @@ TEST(KokkosME, test_pyr5_views)
   );
 }
 
+TEST(KokkosME, test_quad42D_views_shifted_grad_op)
+{
+  test_ME_views<sierra::nalu::AlgTraitsQuad4_2D>(
+    {
+     sierra::nalu::SCS_SHIFTED_GRAD_OP,
+    }
+  );
+}
+
 TEST(KokkosME, test_pyr5_views_shifted_grad_op)
 {
   test_ME_views<sierra::nalu::AlgTraitsPyr5>(
     {
      sierra::nalu::SCS_SHIFTED_GRAD_OP,
+    }
+  );
+}
+
+TEST(KokkosME, test_quad42D_views_gij)
+{
+  test_ME_views<sierra::nalu::AlgTraitsQuad4_2D>(
+    {
+     sierra::nalu::SCS_GRAD_OP,
+     sierra::nalu::SCS_GIJ,
     }
   );
 }
@@ -254,4 +284,3 @@ TEST(KokkosME, test_pyr5_views_gij)
     }
   );
 }
-
